feat: refractor folders and add style.scss done keybord show
