+++ assets/css/style.css
@@ -1,0 +1,74 @@
+span {
+  display: none;
+}
+
+.key {
+  display: flex;
+  width: 40px;
+  height: 40px;
+  justify-content: center;
+  align-items: center;
+  color: #fff;
+  background-color: #444;
+  border-radius: 5px;
+}
+
+.system {
+  background-color: #222;
+}
+
+.keyboard {
+  padding: 0 5px;
+  max-width: 740px;
+  margin: 0 auto;
+  background-color: #888;
+  border: 2px solid #000;
+}
+
+.row {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 5px 0;
+}
+
+#Backspace {
+  width: 100px;
+}
+
+#Tab {
+  width: 60px;
+}
+
+#CapsLock {
+  width: 80px;
+}
+
+#Enter {
+  width: 80px;
+}
+
+#ShiftLeft,
+#ShiftRight {
+  width: 100px;
+}
+
+#Space {
+  width: 306px;
+}
+
+.en .en-norm {
+  display: block;
+}
+
+.shift-en .en-shift {
+  display: block;
+}
+
+.ru .ru-norm {
+  display: block;
+}
+
+.shift-ru .ru-shift {
+  display: block;
+}
