+++ assets/js/keyboard.js
@@ -1,0 +1,833 @@
+const keyboard = {
+  row1: {
+    0: {
+      en: {
+        norm: '`',
+        shift: '~',
+      },
+      ru: {
+        norm: 'ё',
+        shift: 'Ё',
+      },
+      type: 'symbol',
+      code: 'Backquote',
+    },
+    1: {
+      en: {
+        norm: '1',
+        shift: '!',
+      },
+      ru: {
+        norm: '1',
+        shift: '!',
+      },
+
+      type: 'symbol',
+      code: 'Digit1',
+    },
+    2: {
+      en: {
+        norm: '2',
+        shift: '@',
+      },
+      ru: {
+        norm: '2',
+        shift: '"',
+      },
+
+      type: 'symbol',
+      code: 'Digit2',
+    },
+    3: {
+      en: {
+        norm: '3',
+        shift: '#',
+      },
+      ru: {
+        norm: '3',
+        shift: '№',
+      },
+
+      type: 'symbol',
+      code: 'Digit3',
+    },
+    4: {
+      en: {
+        norm: '4',
+        shift: '$',
+      },
+      ru: {
+        norm: '4',
+        shift: ';',
+      },
+
+      type: 'symbol',
+      code: 'Digit4',
+    },
+    5: {
+      en: {
+        norm: '5',
+        shift: '%',
+      },
+      ru: {
+        norm: '5',
+        shift: '%',
+      },
+
+      type: 'symbol',
+      code: 'Digit5',
+    },
+    6: {
+      en: {
+        norm: '6',
+        shift: '^',
+      },
+      ru: {
+        norm: '6',
+        shift: ':',
+      },
+
+      type: 'symbol',
+      code: 'Digit6',
+    },
+    7: {
+      en: {
+        norm: '7',
+        shift: '&',
+      },
+      ru: {
+        norm: '7',
+        shift: '?',
+      },
+
+      type: 'symbol',
+      code: 'Digit7',
+    },
+    8: {
+      en: {
+        norm: '8',
+        shift: '*',
+      },
+      ru: {
+        norm: '8',
+        shift: '*',
+      },
+
+      type: 'symbol',
+      code: 'Digit8',
+    },
+    9: {
+      en: {
+        norm: '9',
+        shift: '(',
+      },
+      ru: {
+        norm: '9',
+        shift: '(',
+      },
+
+      type: 'symbol',
+      code: 'Digit9',
+    },
+    10: {
+      en: {
+        norm: '0',
+        shift: ')',
+      },
+      ru: {
+        norm: '0',
+        shift: ')',
+      },
+
+      type: 'symbol',
+      code: 'Digit0',
+    },
+    11: {
+      en: {
+        norm: '-',
+        shift: '_',
+      },
+      ru: {
+        norm: '-',
+        shift: '_',
+      },
+
+      type: 'symbol',
+      code: 'Minus',
+    },
+    12: {
+      en: {
+        norm: '=',
+        shift: '+',
+      },
+      ru: {
+        norm: '=',
+        shift: '+',
+      },
+
+      type: 'symbol',
+      code: 'Equal',
+    },
+    13: {
+      en: {
+        norm: 'Backspace',
+        shift: 'Backspace',
+      },
+      ru: {
+        norm: 'Backspace',
+        shift: 'Backspace',
+      },
+
+      type: 'system',
+      code: 'Backspace',
+    },
+  },
+  row2: {
+    0: {
+      en: {
+        norm: 'Tab',
+        shift: 'Tab',
+      },
+      ru: {
+        norm: 'Tab',
+        shift: 'Tab',
+      },
+
+      type: 'system',
+      code: 'Tab',
+    },
+    1: {
+      en: {
+        norm: 'q',
+        shift: 'Q',
+      },
+      ru: {
+        norm: 'й',
+        shift: 'Й',
+      },
+
+      type: 'letter',
+      code: 'KeyQ',
+    },
+    2: {
+      en: {
+        norm: 'w',
+        shift: 'W',
+      },
+      ru: {
+        norm: 'ц',
+        shift: 'Ц',
+      },
+
+      type: 'letter',
+      code: 'KeyW',
+    },
+    3: {
+      en: {
+        norm: 'e',
+        shift: 'E',
+      },
+      ru: {
+        norm: 'у',
+        shift: 'У',
+      },
+
+      type: 'letter',
+      code: 'KeyE',
+    },
+    4: {
+      en: {
+        norm: 'r',
+        shift: 'R',
+      },
+      ru: {
+        norm: 'к',
+        shift: 'К',
+      },
+
+      type: 'letter',
+      code: 'KeyR',
+    },
+    5: {
+      en: {
+        norm: 't',
+        shift: 'T',
+      },
+      ru: {
+        norm: 'е',
+        shift: 'Е',
+      },
+
+      type: 'letter',
+      code: 'KeyT',
+    },
+    6: {
+      en: {
+        norm: 'y',
+        shift: 'Y',
+      },
+      ru: {
+        norm: 'н',
+        shift: 'Н',
+      },
+
+      type: 'letter',
+      code: 'KeyY',
+    },
+    7: {
+      en: {
+        norm: 'u',
+        shift: 'U',
+      },
+      ru: {
+        norm: 'г',
+        shift: 'Г',
+      },
+
+      type: 'letter',
+      code: 'KeyU',
+    },
+    8: {
+      en: {
+        norm: 'i',
+        shift: 'I',
+      },
+      ru: {
+        norm: 'ш',
+        shift: 'Ш',
+      },
+
+      type: 'letter',
+      code: 'KeyI',
+    },
+    9: {
+      en: {
+        norm: 'o',
+        shift: 'O',
+      },
+      ru: {
+        norm: 'щ',
+        shift: 'Щ',
+      },
+
+      type: 'letter',
+      code: 'KeyO',
+    },
+    10: {
+      en: {
+        norm: 'p',
+        shift: 'P',
+      },
+      ru: {
+        norm: 'з',
+        shift: 'З',
+      },
+
+      type: 'letter',
+      code: 'KeyP',
+    },
+    11: {
+      en: {
+        norm: '[',
+        shift: '{',
+      },
+      ru: {
+        norm: 'х',
+        shift: 'Х',
+      },
+
+      type: 'symbol',
+      code: 'BracketLeft',
+    },
+    12: {
+      en: {
+        norm: ']',
+        shift: '}',
+      },
+      ru: {
+        norm: 'ъ',
+        shift: 'Ъ',
+      },
+
+      type: 'symbol',
+      code: 'BracketRight',
+    },
+    13: {
+      en: {
+        norm: '\\',
+        shift: '|',
+      },
+      ru: {
+        norm: '\\',
+        shift: '/',
+      },
+
+      type: 'symbol',
+      code: 'Backslash',
+    },
+  },
+  row3: {
+    0: {
+      en: {
+        norm: 'CapsLock',
+        shift: 'CapsLock',
+      },
+      ru: {
+        norm: 'CapsLock',
+        shift: 'CapsLock',
+      },
+
+      type: 'system',
+      code: 'CapsLock',
+    },
+    1: {
+      en: {
+        norm: 'a',
+        shift: 'A',
+      },
+      ru: {
+        norm: 'ф',
+        shift: 'Ф',
+      },
+
+      type: 'letter',
+      code: 'KeyA',
+    },
+    2: {
+      en: {
+        norm: 's',
+        shift: 'S',
+      },
+      ru: {
+        norm: 'ы',
+        shift: 'Ы',
+      },
+
+      type: 'letter',
+      code: 'KeyS',
+    },
+    3: {
+      en: {
+        norm: 'd',
+        shift: 'D',
+      },
+      ru: {
+        norm: 'в',
+        shift: 'В',
+      },
+
+      type: 'letter',
+      code: 'KeyD',
+    },
+    4: {
+      en: {
+        norm: 'f',
+        shift: 'F',
+      },
+      ru: {
+        norm: 'а',
+        shift: 'А',
+      },
+
+      type: 'letter',
+      code: 'KeyF',
+    },
+    5: {
+      en: {
+        norm: 'g',
+        shift: 'G',
+      },
+      ru: {
+        norm: 'п',
+        shift: 'П',
+      },
+
+      type: 'letter',
+      code: 'KeyG',
+    },
+    6: {
+      en: {
+        norm: 'h',
+        shift: 'G',
+      },
+      ru: {
+        norm: 'р',
+        shift: 'Р',
+      },
+
+      type: 'letter',
+      code: 'KeyH',
+    },
+    7: {
+      en: {
+        norm: 'j',
+        shift: 'J',
+      },
+      ru: {
+        norm: 'о',
+        shift: 'О',
+      },
+
+      type: 'letter',
+      code: 'KeyJ',
+    },
+    8: {
+      en: {
+        norm: 'k',
+        shift: 'K',
+      },
+      ru: {
+        norm: 'л',
+        shift: 'Л',
+      },
+
+      type: 'letter',
+      code: 'KeyK',
+    },
+    9: {
+      en: {
+        norm: 'l',
+        shift: 'L',
+      },
+      ru: {
+        norm: 'д',
+        shift: 'Д',
+      },
+
+      type: 'letter',
+      code: 'KeyL',
+    },
+    10: {
+      en: {
+        norm: ';',
+        shift: ':',
+      },
+      ru: {
+        norm: 'ж',
+        shift: 'Ж',
+      },
+
+      type: 'symbol',
+      code: 'Semicolon',
+    },
+    11: {
+      en: {
+        norm: '\'',
+        shift: '"',
+      },
+      ru: {
+        norm: 'э',
+        shift: 'Э',
+      },
+
+      type: 'symbol',
+      code: 'Quote',
+    },
+    12: {
+      en: {
+        norm: 'Enter',
+        shift: 'Enter',
+      },
+      ru: {
+        norm: 'Enter',
+        shift: 'Enter',
+      },
+
+      type: 'system',
+      code: 'Enter',
+    },
+  },
+  row4: {
+    0: {
+      en: {
+        norm: 'Shift',
+        shift: 'Shift',
+      },
+      ru: {
+        norm: 'Shift',
+        shift: 'Shift',
+      },
+
+      type: 'system',
+      code: 'ShiftLeft',
+    },
+    1: {
+      en: {
+        norm: 'z',
+        shift: 'Z',
+      },
+      ru: {
+        norm: 'я',
+        shift: 'Я',
+      },
+
+      type: 'letter',
+      code: 'KeyZ',
+    },
+    2: {
+      en: {
+        norm: 'x',
+        shift: 'X',
+      },
+      ru: {
+        norm: 'ч',
+        shift: 'Ч',
+      },
+
+      type: 'letter',
+      code: 'KeyX',
+    },
+    3: {
+      en: {
+        norm: 'c',
+        shift: 'C',
+      },
+      ru: {
+        norm: 'с',
+        shift: 'С',
+      },
+
+      type: 'letter',
+      code: 'KeyC',
+    },
+    4: {
+      en: {
+        norm: 'v',
+        shift: 'V',
+      },
+      ru: {
+        norm: 'м',
+        shift: 'М',
+      },
+
+      type: 'letter',
+      code: 'KeyV',
+    },
+    5: {
+      en: {
+        norm: 'b',
+        shift: 'B',
+      },
+      ru: {
+        norm: 'и',
+        shift: 'И',
+      },
+
+      type: 'letter',
+      code: 'KeyB',
+    },
+    6: {
+      en: {
+        norm: 'n',
+        shift: 'N',
+      },
+      ru: {
+        norm: 'т',
+        shift: 'Т',
+      },
+
+      type: 'letter',
+      code: 'KeyN',
+    },
+    7: {
+      en: {
+        norm: 'm',
+        shift: 'M',
+      },
+      ru: {
+        norm: 'ь',
+        shift: 'Ь',
+      },
+
+      type: 'letter',
+      code: 'KeyM',
+    },
+    8: {
+      en: {
+        norm: ',',
+        shift: '<',
+      },
+      ru: {
+        norm: 'б',
+        shift: 'Б',
+      },
+
+      type: 'symbol',
+      code: 'Comma',
+    },
+    9: {
+      en: {
+        norm: '.',
+        shift: '>',
+      },
+      ru: {
+        norm: 'ю',
+        shift: 'Ю',
+      },
+
+      type: 'symbol',
+      code: 'Period',
+    },
+    10: {
+      en: {
+        norm: '/',
+        shift: '?',
+      },
+      ru: {
+        norm: '.',
+        shift: ',',
+      },
+
+      type: 'symbol',
+      code: 'Slash',
+    },
+    11: {
+      en: {
+        norm: '↑',
+        shift: '↑"',
+      },
+      ru: {
+        norm: '↑',
+        shift: '↑"',
+      },
+
+      type: 'system',
+      code: 'ArrowUp',
+    },
+    12: {
+      en: {
+        norm: 'Shift',
+        shift: 'Shift',
+      },
+      ru: {
+        norm: 'Shift',
+        shift: 'Shift',
+      },
+
+      type: 'system',
+      code: 'ShiftRight',
+    },
+  },
+
+  row5: {
+    0: {
+      en: {
+        norm: 'Ctrl',
+        shift: 'Ctrl',
+      },
+      ru: {
+        norm: 'Ctrl',
+        shift: 'Ctrl',
+      },
+
+      type: 'system',
+      code: 'ControlLeft',
+    },
+    1: {
+      en: {
+        norm: 'Win',
+        shift: 'Win',
+      },
+      ru: {
+        norm: 'Win',
+        shift: 'Win',
+      },
+
+      type: 'system',
+      code: 'MetaLeft',
+    },
+    2: {
+      en: {
+        norm: 'Alt',
+        shift: 'Alt',
+      },
+      ru: {
+        norm: 'Alt',
+        shift: 'Alt',
+      },
+
+      type: 'system',
+      code: 'AltLeft',
+    },
+    3: {
+      en: {
+        norm: 'Space',
+        shift: 'Space',
+      },
+      ru: {
+        norm: 'Space',
+        shift: 'Space',
+      },
+
+      type: 'system',
+      code: 'Space',
+    },
+    4: {
+      en: {
+        norm: 'Alt',
+        shift: 'Alt',
+      },
+      ru: {
+        norm: 'Alt',
+        shift: 'Alt',
+      },
+
+      type: 'system',
+      code: 'AltRight',
+    },
+    5: {
+      en: {
+        norm: '←',
+        shift: '←',
+      },
+      ru: {
+        norm: '←',
+        shift: '←',
+      },
+
+      type: 'system',
+      code: 'ArrowLeft',
+    },
+    6: {
+      en: {
+        norm: '↓',
+        shift: '↓',
+      },
+      ru: {
+        norm: '↓',
+        shift: '↓',
+      },
+
+      type: 'system',
+      code: 'ArrowDown',
+    },
+    7: {
+      en: {
+        norm: '→',
+        shift: '→',
+      },
+      ru: {
+        norm: '→',
+        shift: '→',
+      },
+
+      type: 'system',
+      code: 'ArrowRight',
+    },
+    8: {
+      en: {
+        norm: 'Ctrl',
+        shift: 'Ctrl',
+      },
+      ru: {
+        norm: 'Ctrl',
+        shift: 'Ctrl',
+      },
+
+      type: 'system',
+      code: 'ControlRight',
+    },
+  },
+};
+
+export default keyboard;
