--- js/keyboard.js
@@ -1,786 +1,0 @@
-const keyboard = {
-  en: {
-    row1: {
-      0: {
-        norm: '`',
-        shift: '~',
-        type: 'symbol',
-        code: 'Backquote',
-      },
-      1: {
-        norm: '1',
-        shift: '!',
-        type: 'symbol',
-        code: 'Digit1',
-      },
-      2: {
-        norm: '2',
-        shift: '@',
-        type: 'symbol',
-        code: 'Digit2',
-      },
-      3: {
-        norm: '3',
-        shift: '#',
-        type: 'symbol',
-        code: 'Digit3',
-      },
-      4: {
-        norm: '4',
-        shift: '$',
-        type: 'symbol',
-        code: 'Digit4',
-      },
-      5: {
-        norm: '5',
-        shift: '%',
-        type: 'symbol',
-        code: 'Digit5',
-      },
-      6: {
-        norm: '6',
-        shift: '^',
-        type: 'symbol',
-        code: 'Digit6',
-      },
-      7: {
-        norm: '7',
-        shift: '&',
-        type: 'symbol',
-        code: 'Digit7',
-      },
-      8: {
-        norm: '8',
-        shift: '*',
-        type: 'symbol',
-        code: 'Digit8',
-      },
-      9: {
-        norm: '9',
-        shift: '(',
-        type: 'symbol',
-        code: 'Digit9',
-      },
-      10: {
-        norm: '0',
-        shift: ')',
-        type: 'symbol',
-        code: 'Digit0',
-      },
-      11: {
-        norm: '-',
-        shift: '_',
-        type: 'symbol',
-        code: 'Minus',
-      },
-      12: {
-        norm: '=',
-        shift: '+',
-        type: 'symbol',
-        code: 'Equal',
-      },
-      13: {
-        norm: 'Backspace',
-        shift: 'Backspace',
-        type: 'system',
-        code: 'Backspace',
-      },
-    },
-    row2: {
-      0: {
-        norm: 'Tab',
-        shift: 'Tab',
-        type: 'system',
-        code: 'Tab',
-      },
-      1: {
-        norm: 'q',
-        shift: 'Q',
-        type: 'letter',
-        code: 'KeyQ',
-      },
-      2: {
-        norm: 'w',
-        shift: 'W',
-        type: 'letter',
-        code: 'KeyW',
-      },
-      3: {
-        norm: 'e',
-        shift: 'E',
-        type: 'letter',
-        code: 'KeyE',
-      },
-      4: {
-        norm: 'r',
-        shift: 'R',
-        type: 'letter',
-        code: 'KeyR',
-      },
-      5: {
-        norm: 't',
-        shift: 'T',
-        type: 'letter',
-        code: 'KeyT',
-      },
-      6: {
-        norm: 'y',
-        shift: 'Y',
-        type: 'letter',
-        code: 'KeyY',
-      },
-      7: {
-        norm: 'u',
-        shift: 'U',
-        type: 'letter',
-        code: 'KeyU',
-      },
-      8: {
-        norm: 'i',
-        shift: 'I',
-        type: 'letter',
-        code: 'KeyI',
-      },
-      9: {
-        norm: 'o',
-        shift: 'O',
-        type: 'letter',
-        code: 'KeyO',
-      },
-      10: {
-        norm: 'p',
-        shift: 'P',
-        type: 'letter',
-        code: 'KeyP',
-      },
-      11: {
-        norm: '[',
-        shift: '{',
-        type: 'symbol',
-        code: 'BracketLeft',
-      },
-      12: {
-        norm: ']',
-        shift: '}',
-        type: 'symbol',
-        code: 'BracketRight',
-      },
-      13: {
-        norm: '\\',
-        shift: '|',
-        type: 'symbol',
-        code: 'Backslash',
-      },
-    },
-    row3: {
-      0: {
-        norm: 'CapsLock',
-        shift: 'CapsLock',
-        type: 'system',
-        code: 'CapsLock',
-      },
-      1: {
-        norm: 'a',
-        shift: 'A',
-        type: 'letter',
-        code: 'KeyA',
-      },
-      2: {
-        norm: 's',
-        shift: 'S',
-        type: 'letter',
-        code: 'KeyS',
-      },
-      3: {
-        norm: 'd',
-        shift: 'D',
-        type: 'letter',
-        code: 'KeyD',
-      },
-      4: {
-        norm: 'f',
-        shift: 'F',
-        type: 'letter',
-        code: 'KeyF',
-      },
-      5: {
-        norm: 'g',
-        shift: 'G',
-        type: 'letter',
-        code: 'KeyG',
-      },
-      6: {
-        norm: 'h',
-        shift: 'G',
-        type: 'letter',
-        code: 'KeyH',
-      },
-      7: {
-        norm: 'j',
-        shift: 'J',
-        type: 'letter',
-        code: 'KeyJ',
-      },
-      8: {
-        norm: 'k',
-        shift: 'K',
-        type: 'letter',
-        code: 'KeyK',
-      },
-      9: {
-        norm: 'l',
-        shift: 'L',
-        type: 'letter',
-        code: 'KeyL',
-      },
-      10: {
-        norm: ';',
-        shift: ':',
-        type: 'symbol',
-        code: 'Semicolon',
-      },
-      11: {
-        norm: '\'',
-        shift: '"',
-        type: 'symbol',
-        code: 'Quote',
-      },
-      12: {
-        norm: 'Enter',
-        shift: 'Enter',
-        type: 'system',
-        code: 'Enter',
-      },
-    },
-    row4: {
-      0: {
-        norm: 'Shift',
-        shift: 'Shift',
-        type: 'system',
-        code: 'ShiftLeft',
-      },
-      1: {
-        norm: 'z',
-        shift: 'Z',
-        type: 'letter',
-        code: 'KeyZ',
-      },
-      2: {
-        norm: 'x',
-        shift: 'X',
-        type: 'letter',
-        code: 'KeyX',
-      },
-      3: {
-        norm: 'c',
-        shift: 'C',
-        type: 'letter',
-        code: 'KeyC',
-      },
-      4: {
-        norm: 'v',
-        shift: 'V',
-        type: 'letter',
-        code: 'KeyV',
-      },
-      5: {
-        norm: 'b',
-        shift: 'B',
-        type: 'letter',
-        code: 'KeyB',
-      },
-      6: {
-        norm: 'n',
-        shift: 'N',
-        type: 'letter',
-        code: 'KeyN',
-      },
-      7: {
-        norm: 'm',
-        shift: 'M',
-        type: 'letter',
-        code: 'KeyM',
-      },
-      8: {
-        norm: ',',
-        shift: '<',
-        type: 'symbol',
-        code: 'Comma',
-      },
-      9: {
-        norm: '.',
-        shift: '>',
-        type: 'symbol',
-        code: 'Period',
-      },
-      10: {
-        norm: '/',
-        shift: '?',
-        type: 'symbol',
-        code: 'Slash',
-      },
-      11: {
-        norm: '↑',
-        shift: '↑"',
-        type: 'system',
-        code: 'ArrowUp',
-      },
-      12: {
-        norm: 'Shift',
-        shift: 'Shift',
-        type: 'system',
-        code: 'ShiftRight',
-      },
-    },
-
-    row5: {
-      0: {
-        norm: 'Ctrl',
-        shift: 'Ctrl',
-        type: 'system',
-        code: 'ControlLeft',
-      },
-      1: {
-        norm: 'Win',
-        shift: 'Win',
-        type: 'system',
-        code: 'MetaLeft',
-      },
-      2: {
-        norm: 'Alt',
-        shift: 'Alt',
-        type: 'system',
-        code: 'AltLeft',
-      },
-      3: {
-        norm: 'Space',
-        shift: 'Space',
-        type: 'system',
-        code: 'Space',
-      },
-      4: {
-        norm: 'Alt',
-        shift: 'Alt',
-        type: 'system',
-        code: 'AltRight',
-      },
-      5: {
-        norm: '←',
-        shift: '←',
-        type: 'system',
-        code: 'ArrowLeft',
-      },
-      6: {
-        norm: '↓',
-        shift: '↓',
-        type: 'system',
-        code: 'ArrowDown',
-      },
-      7: {
-        norm: '→',
-        shift: '→',
-        type: 'system',
-        code: 'ArrowRight',
-      },
-      8: {
-        norm: 'Ctrl',
-        shift: 'Ctrl',
-        type: 'system',
-        code: 'ControlRight',
-      },
-    },
-  },
-  ru: {
-    row1: {
-      0: {
-        norm: 'ё',
-        shift: 'Ё',
-        type: 'letter',
-        code: 'Backquote',
-      },
-      1: {
-        norm: '1',
-        shift: '!',
-        type: 'symbol',
-        code: 'Digit1',
-      },
-      2: {
-        norm: '2',
-        shift: '"',
-        type: 'symbol',
-        code: 'Digit2',
-      },
-      3: {
-        norm: '3',
-        shift: '№',
-        type: 'symbol',
-        code: 'Digit3',
-      },
-      4: {
-        norm: '4',
-        shift: ';',
-        type: 'symbol',
-        code: 'Digit4',
-      },
-      5: {
-        norm: '5',
-        shift: '%',
-        type: 'symbol',
-        code: 'Digit5',
-      },
-      6: {
-        norm: '6',
-        shift: ':',
-        type: 'symbol',
-        code: 'Digit6',
-      },
-      7: {
-        norm: '7',
-        shift: '?',
-        type: 'symbol',
-        code: 'Digit7',
-      },
-      8: {
-        norm: '8',
-        shift: '*',
-        type: 'symbol',
-        code: 'Digit8',
-      },
-      9: {
-        norm: '9',
-        shift: '(',
-        type: 'symbol',
-        code: 'Digit9',
-      },
-      10: {
-        norm: '0',
-        shift: ')',
-        type: 'symbol',
-        code: 'Digit0',
-      },
-      11: {
-        norm: '-',
-        shift: '_',
-        type: 'symbol',
-        code: 'Minus',
-      },
-      12: {
-        norm: '=',
-        shift: '+',
-        type: 'symbol',
-        code: 'Equal',
-      },
-      13: {
-        norm: 'Backspace',
-        shift: 'Backspace',
-        type: 'system',
-        code: 'Backspace',
-      },
-    },
-    row2: {
-      0: {
-        norm: 'Tab',
-        shift: 'Tab',
-        type: 'system',
-        code: 'Tab',
-      },
-      1: {
-        norm: 'й',
-        shift: 'Й',
-        type: 'letter',
-        code: 'KeyQ',
-      },
-      2: {
-        norm: 'ц',
-        shift: 'Ц',
-        type: 'letter',
-        code: 'KeyW',
-      },
-      3: {
-        norm: 'у',
-        shift: 'У',
-        type: 'letter',
-        code: 'KeyE',
-      },
-      4: {
-        norm: 'к',
-        shift: 'К',
-        type: 'letter',
-        code: 'KeyR',
-      },
-      5: {
-        norm: 'е',
-        shift: 'Е',
-        type: 'letter',
-        code: 'KeyT',
-      },
-      6: {
-        norm: 'н',
-        shift: 'Н',
-        type: 'letter',
-        code: 'KeyY',
-      },
-      7: {
-        norm: 'г',
-        shift: 'Г',
-        type: 'letter',
-        code: 'KeyU',
-      },
-      8: {
-        norm: 'ш',
-        shift: 'Ш',
-        type: 'letter',
-        code: 'KeyI',
-      },
-      9: {
-        norm: 'щ',
-        shift: 'Щ',
-        type: 'letter',
-        code: 'KeyO',
-      },
-      10: {
-        norm: 'з',
-        shift: 'З',
-        type: 'letter',
-        code: 'KeyP',
-      },
-      11: {
-        norm: 'х',
-        shift: 'Х',
-        type: 'letter',
-        code: 'BracketLeft',
-      },
-      12: {
-        norm: 'ъ',
-        shift: 'Ъ',
-        type: 'letter',
-        code: 'BracketRight',
-      },
-      13: {
-        norm: '\\',
-        shift: '/',
-        type: 'symbol',
-        code: 'Backslash',
-      },
-    },
-    row3: {
-      0: {
-        norm: 'CapsLock',
-        shift: 'CapsLock',
-        type: 'system',
-        code: 'CapsLock',
-      },
-      1: {
-        norm: 'ф',
-        shift: 'Ф',
-        type: 'letter',
-        code: 'KeyA',
-      },
-      2: {
-        norm: 'ы',
-        shift: 'Ы',
-        type: 'letter',
-        code: 'KeyS',
-      },
-      3: {
-        norm: 'в',
-        shift: 'В',
-        type: 'letter',
-        code: 'KeyD',
-      },
-      4: {
-        norm: 'а',
-        shift: 'А',
-        type: 'letter',
-        code: 'KeyF',
-      },
-      5: {
-        norm: 'п',
-        shift: 'П',
-        type: 'letter',
-        code: 'KeyG',
-      },
-      6: {
-        norm: 'р',
-        shift: 'Р',
-        type: 'letter',
-        code: 'KeyH',
-      },
-      7: {
-        norm: 'о',
-        shift: 'О',
-        type: 'letter',
-        code: 'KeyJ',
-      },
-      8: {
-        norm: 'л',
-        shift: 'Л',
-        type: 'letter',
-        code: 'KeyK',
-      },
-      9: {
-        norm: 'д',
-        shift: 'Д',
-        type: 'letter',
-        code: 'KeyL',
-      },
-      10: {
-        norm: 'ж',
-        shift: 'Ж',
-        type: 'letter',
-        code: 'Semicolon',
-      },
-      11: {
-        norm: 'э',
-        shift: 'Э',
-        type: 'letter',
-        code: 'Quote',
-      },
-      12: {
-        norm: 'Enter',
-        shift: 'Enter',
-        type: 'system',
-        code: 'Enter',
-      },
-    },
-    row4: {
-      0: {
-        norm: 'Shift',
-        shift: 'Shift',
-        type: 'system',
-        code: 'ShiftLeft',
-      },
-      1: {
-        norm: 'я',
-        shift: 'Я',
-        type: 'letter',
-        code: 'KeyZ',
-      },
-      2: {
-        norm: 'ч',
-        shift: 'Ч',
-        type: 'letter',
-        code: 'KeyX',
-      },
-      3: {
-        norm: 'с',
-        shift: 'С',
-        type: 'letter',
-        code: 'KeyC',
-      },
-      4: {
-        norm: 'м',
-        shift: 'М',
-        type: 'letter',
-        code: 'KeyV',
-      },
-      5: {
-        norm: 'и',
-        shift: 'И',
-        type: 'letter',
-        code: 'KeyB',
-      },
-      6: {
-        norm: 'т',
-        shift: 'Т',
-        type: 'letter',
-        code: 'KeyN',
-      },
-      7: {
-        norm: 'ь',
-        shift: 'Ь',
-        type: 'letter',
-        code: 'KeyM',
-      },
-      8: {
-        norm: 'б',
-        shift: 'Б',
-        type: 'symbol',
-        code: 'Comma',
-      },
-      9: {
-        norm: 'ю',
-        shift: 'Ю',
-        type: 'symbol',
-        code: 'Period',
-      },
-      10: {
-        norm: '.',
-        shift: ',',
-        type: 'symbol',
-        code: 'Slash',
-      },
-      11: {
-        norm: '↑',
-        shift: '↑"',
-        type: 'system',
-        code: 'ArrowUp',
-      },
-      12: {
-        norm: 'Shift',
-        shift: 'Shift',
-        type: 'system',
-        code: 'ShiftRight',
-      },
-    },
-
-    row5: {
-      0: {
-        norm: 'Ctrl',
-        shift: 'Ctrl',
-        type: 'system',
-        code: 'ControlLeft',
-      },
-      1: {
-        norm: 'Win',
-        shift: 'Win',
-        type: 'system',
-        code: 'MetaLeft',
-      },
-      2: {
-        norm: 'Alt',
-        shift: 'Alt',
-        type: 'system',
-        code: 'AltLeft',
-      },
-      3: {
-        norm: 'Space',
-        shift: 'Space',
-        type: 'system',
-        code: 'Space',
-      },
-      4: {
-        norm: 'Alt',
-        shift: 'Alt',
-        type: 'system',
-        code: 'AltRight',
-      },
-      5: {
-        norm: '←',
-        shift: '←',
-        type: 'system',
-        code: 'ArrowLeft',
-      },
-      6: {
-        norm: '↓',
-        shift: '↓',
-        type: 'system',
-        code: 'ArrowDown',
-      },
-      7: {
-        norm: '→',
-        shift: '→',
-        type: 'system',
-        code: 'ArrowRight',
-      },
-      8: {
-        norm: 'Ctrl',
-        shift: 'Ctrl',
-        type: 'system',
-        code: 'ControlRight',
-      },
-    },
-  },
-};
-
-export default keyboard;
