--- style.css
@@ -1,3 +1,0 @@
-span {
-	display: none;
-}
